Move Audit client
