--- src/main/java/uk/gov/digital/ho/hocs/casework/auditClient/EventType.java
@@ -1,10 +1,0 @@
-package uk.gov.digital.ho.hocs.casework.auditClient;
-
-import lombok.Getter;
-
-public enum EventType {
-    CASE_ALLOCATED_SELF,
-    CASE_ALLOCATED_TO,
-    CASE_CREATED,
-    CASE_VIEWED;
-}
