+++ src/main/java/uk/gov/digital/ho/hocs/casework/client/auditclient/AuditClient.java
@@ -1,4 +1,4 @@
-package uk.gov.digital.ho.hocs.casework.auditclient;
+package uk.gov.digital.ho.hocs.casework.client.auditclient;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.extern.slf4j.Slf4j;
@@ -7,7 +7,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 import uk.gov.digital.ho.hocs.casework.application.RequestData;
-import uk.gov.digital.ho.hocs.casework.auditclient.dto.CreateAuditRequest;
+import uk.gov.digital.ho.hocs.casework.client.auditclient.dto.CreateAuditRequest;
 import uk.gov.digital.ho.hocs.casework.domain.model.CaseData;
 
 import java.time.LocalDateTime;
